Checkstyle compliance
