+++ gradle.properties
@@ -1,17 +1,15 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
-
 # Fabric Properties
-	# check these on https://fabricmc.net/use
-	minecraft_version=1.15.2
-	yarn_build=14
-	loader_version=0.7.8+build.184
-
+# check these on https://fabricmc.net/use
+minecraft_version=1.15.2
+yarn_build=14
+loader_version=0.7.8+build.184
 # Mod Properties
-	mod_version = 1.1.1
-	maven_group = io.github.kvverti
-	library_name = banner-pattern-api
-	library_version = 1
+mod_version=1.1.1
+maven_group=io.github.fablabsmc
+library_name=banner-pattern-api
+library_version=1
 # Dependencies
-	# currently not on the main fabric site, check on the maven: https://maven.fabricmc.net/net/fabricmc/fabric-api/
-	fabric_version=0.4.20+build.273-1.15
+# currently not on the main fabric site, check on the maven: https://maven.fabricmc.net/net/fabricmc/fabric-api/
+fabric_version=0.4.20+build.273-1.15
