+++ src/main/resources/bannerpp.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "io.github.kvverti.bannerpp.mixin",
+  "package": "io.github.fablabsmc.fablabs.mixin.bannerpattern",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "BannerBlockEntityMixin",
