+++ src/main/resources/fabric.mod.json
@@ -20,10 +20,10 @@
   "environment": "*",
   "entrypoints": {
     "main": [
-      "io.github.kvverti.bannerpp.impl.Bannerpp"
+      "io.github.fablabsmc.fablabs.impl.bannerpattern.Bannerpp"
     ],
     "client": [
-      "io.github.kvverti.bannerpp.impl.BannerppClient"
+      "io.github.fablabsmc.fablabs.impl.bannerpattern.BannerppClient"
     ]
   },
   "mixins": [
